requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pywebview==4.4.1
jinja2==3.1.3
aiofiles==23.2.1
